requirements.txt
PyQt5>=5.15.0
pandas>=1.3
pandas-ta>=0.3.14
mplfinance>=0.12.9
websocket-client>=1.6
requests>=2.25